#251 move tiering interfaces to core module + cleanup previous prototype of caching store
