+++ impl/src/main/java/org/ehcache/internal/store/disk/DiskStore.java
@@ -32,8 +32,8 @@ import org.ehcache.function.Predicates;
 import org.ehcache.internal.SystemTimeSource;
 import org.ehcache.internal.TimeSource;
 import org.ehcache.internal.TimeSourceConfiguration;
-import org.ehcache.internal.store.tiering.AuthoritativeTier;
-import org.ehcache.internal.store.tiering.CachingTier;
+import org.ehcache.spi.cache.tiering.AuthoritativeTier;
+import org.ehcache.spi.cache.tiering.CachingTier;
 import org.ehcache.internal.store.disk.DiskStorageFactory.Element;
 import org.ehcache.spi.ServiceProvider;
 import org.ehcache.spi.cache.Store;
